+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.board;
 
+import chess.domain.piece.Color;
 import java.util.ArrayList;
 import java.util.List;
