+++ src/main/java/chess/domain/board/FileCoordinate.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 public enum FileCoordinate {
     A,
