+++ src/main/java/chess/domain/board/RankFactory.java
@@ -1,5 +1,8 @@
-package chess.domain;
+package chess.domain.board;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
 import java.util.ArrayList;
 import java.util.List;
