+++ src/main/java/chess/domain/board/RankType.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 public enum RankType {
     SIDE_RANK,
