+++ src/main/java/chess/domain/board/Square.java
@@ -1,4 +1,6 @@
-package chess.domain;
+package chess.domain.board;
+
+import chess.domain.piece.Piece;
 
 public class Square {
