+++ src/main/java/chess/domain/piece/Color.java
@@ -1,4 +1,6 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.board.RankCoordinate;
 
 public enum Color {
     BLACK,
