+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Piece {
