+++ src/test/java/chess/domain/piece/ColorTest.java
@@ -1,8 +1,9 @@
-package chess.domain;
+package chess.domain.piece;
 
-import static chess.domain.Color.of;
+import static chess.domain.piece.Color.of;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.board.RankCoordinate;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.params.ParameterizedTest;
